Fix human sweep command to work with troop lift
Previously, the we assumed that all guerrillas in sweep
spaces were activated simultaneously and thus did not allow
a special activity (eg. Troop Lift) to occur between activations.

Now, we allow the user to select the sweep spaces to be activated
one by one and allow a special activity to occur in between.

This allows the same troop cubes to activate guerrillas in more
than one space.