fix: show human ops menu even if no valid spaces
For assault and sweep if there were no valid spaces
we we showing a message and aborting the human turn.
But, if the user has chosen to include a special
activity, then this was precluding them from creating
a valid space via the special activity.  So I removed
the checks.